Update all scripts
