+++ Scripts/MCF.py
@@ -2,27 +2,14 @@
 # coding: utf-8
 
 # core
-import sys, os, re, argparse
-import numpy as np
+import sys, argparse
 from datetime import datetime
-import pandas as pd
-from scipy import spatial
-from tqdm import tqdm
-from multiprocessing import Pool, set_start_method
-
-# plotting
-import matplotlib.pyplot as plt
 # image recognition
 import cv2
-import imutils
-import pytesseract
-from pytesseract import image_to_string
-import skimage.measure
-from skimage import io, data
-from skimage.util import img_as_float, img_as_ubyte
 from pathlib import Path
 # Custom utilities
 from utilities import *
+from imagetools import *
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -48,18 +35,15 @@ def main():
     parser.add_argument("--prefix", type=str,
                     action="store", dest="prefix",
                     help="New prefix for output files. By default the new files will reflect the input file's basename")
-    parser.add_argument("--sliding-window", type=bool, default=True,
-                    action="store", dest="sliding_window",
-                    help="Use sliding window approach. Default=True")
-    parser.add_argument("--stepsize", type=int,
-                    action="store", dest="stepsize",
-                    help="Slide step size in pixels (same in x and y directions). Default=500")
-    parser.add_argument("--winW", type=int,
-                    action="store", dest="winW",
-                    help="Window width in pixels. Default=1000")
-    parser.add_argument("--winH", type=int,
-                    action="store", dest="winH",
-                    help="Window height in pixels. Default=1000")
+    # parser.add_argument("--stepsize", type=int,
+    #                 action="store", dest="stepsize",
+    #                 help="Slide step size in pixels (same in x and y directions). Default=500")
+    # parser.add_argument("--winW", type=int,
+    #                 action="store", dest="winW",
+    #                 help="Window width in pixels. Default=1000")
+    # parser.add_argument("--winH", type=int,
+    #                 action="store", dest="winH",
+    #                 help="Window height in pixels. Default=1000")
     parser.add_argument("--neighborhood", type=int, default=10,
                     action="store", dest="neighborhood",
                     help="Neighborhood size in pixels determining a unique contour. Default=10")
@@ -108,28 +92,14 @@ def main():
     kwargs['k_gradient'] = args.k_gradient
     kwargs['k_foreground'] = args.k_foreground
 
-    h, w, c = cv2.imread(args.input).shape
-    if not args.winW:
-        winW = round((w / 5) / 100) * 100
-    else:
-        winW = args.winW
-    if not args.winH:
-        winH = round((h / 5) / 100) * 100
-    else:
-        winH = args.winH
-    if not args.stepsize:
-        stepsize = int(min([winH, winW])/2)
-    else:
-        stepsize = args.stepsize
-
     if not args.prefix:
         prefix = Path(args.input).stem
     else:
         prefix = args.prefix
 
 
-    print("[{}] MCF command:\n\t python MCF.py --prefix {} --sliding-window {} --stepsize {} --winW {} --winH {} --neighborhood {} --dpi {} --debug {} --k_blur {} --C {} --blocksize {} --k_laplacian {} --k_dilate {} --k_gradient {} --k_foreground {} --Amin {} --Amax {:d}".format(
-        datetime.now().strftime('%d %b %Y %H:%M:%S'), args.prefix, args.sliding_window, stepsize, winW, winH, args.neighborhood, args.dpi, args.debug, args.k_blur, args.C, args.blocksize, args.k_laplacian, args.k_dilate, args.k_gradient, args.k_foreground, int(args.Amin), int(args.Amax)))
+    print("[{}] MCF command:\n\t python MCF.py --prefix {} --neighborhood {} --dpi {} --debug {} --k_blur {} --C {} --blocksize {} --k_laplacian {} --k_dilate {} --k_gradient {} --k_foreground {} --Amin {} --Amax {:d}".format(
+        datetime.now().strftime('%d %b %Y %H:%M:%S'), args.prefix, args.neighborhood, args.dpi, args.debug, args.k_blur, args.C, args.blocksize, args.k_laplacian, args.k_dilate, args.k_gradient, args.k_foreground, int(args.Amin), int(args.Amax)))
 
     if args.input.startswith("~/"):
         input_path = Path(Path.home() / args.input[2:])
@@ -143,7 +113,7 @@ def main():
     print("[{}] Input file: {}".format(datetime.now().strftime('%d %b %Y %H:%M:%S'), input_path.absolute()))
     print("[{}] Output directory: {}".format(datetime.now().strftime('%d %b %Y %H:%M:%S'), output_dir_path.absolute()))
 
-    suffixes = ["summary.csv", "contour_data.csv", "contour_data.pkl", "tif", "noindex.tif"]
+    suffixes = ["contour_data.csv", "contour_data.json", "tif", "noindex.tif"]
     conflicts = [output_dir_path / "{}.{}".format(prefix, suffix) for suffix in suffixes]
 
     if any([c.exists() for c in conflicts]):
@@ -155,11 +125,10 @@ def main():
             print("\tDelete output files or change output prefix to rerun analysis")
             sys.exit()
 
-    process_image(image_path=input_path, prefix=prefix, sliding_window=args.sliding_window, stepsize=stepsize, winW=winW, winH=winH, Amin=args.Amin, Amax=args.Amax, neighborhood=args.neighborhood, output_dir=output_dir_path, dpi=args.dpi, debug=args.debug, **kwargs)
+    process_image(image_path=input_path, prefix=prefix, Amin=args.Amin, Amax=args.Amax, neighborhood=args.neighborhood, output_dir=output_dir_path, dpi=args.dpi, debug=args.debug, **kwargs)
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 if __name__ == "__main__":
-    set_start_method("spawn")
     start = datetime.now()
     main()
     end = datetime.now()
